+++ objectified-ui/public/WHATS_NEW.md
@@ -63,6 +63,9 @@ We continue to improve the platform based on your feedback with improvements and
   - Added smart guidelines when dragging nodes for better alignment
   - Added spacing utility for multiple nodes (selected with Shift key)
   - Added simple automatic layout algorithm for arranging nodes (needs further refinement)
+  - Added edge styling:
+    - Styling for direct, optional, weak, and bidirectional relationships
+    - 
 - Added property templates
   - Fixed property copying to a class that contains an object with nested properties
 - Dashboard improvements